+++ src/components/ui/filters/week-selector.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useRouter } from 'next/navigation'
+import { useRouter, useSearchParams } from 'next/navigation'
 
 interface WeekSelectorProps {
   weeks: { label: string, value: string }[];
@@ -8,10 +8,21 @@ interface WeekSelectorProps {
 
 export function WeekSelector({ weeks }: WeekSelectorProps) {
   const router = useRouter();
+  const searchParams = useSearchParams();
+  const currentWeek = searchParams.get('week');
+
+  const handleParamChange = (paramName: string, paramValue: string, path: string) => {
+    const params = new URLSearchParams(searchParams.toString());
+    if (paramValue) {
+      params.set(paramName, paramValue);
+    } else {
+      params.delete(paramName);
+    }
+    router.push(`${path}?${params.toString()}`);
+  };
 
   const handleWeekChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    const selectedWeek = event.target.value;
-    router.push(`/hours-worked/${selectedWeek}`);
+    handleParamChange('week', event.target.value, '/hours-worked');
   };
 
   return (
@@ -19,7 +30,7 @@ export function WeekSelector({ weeks }: WeekSelectorProps) {
       <label htmlFor="week-select" className="mr-2">Filter by week:</label>
       <select
         id="week-select"
-        defaultValue={weeks[weeks.length - 1].value}
+        value={currentWeek || ''}
         onChange={handleWeekChange}
         className="border p-1 rounded"
       >
